Replaced UserControl with SplitContainer and toggle
